move hostname hack to v37
netbox v24 does not work have __swi filter query
